Fixed integer overflow.

+++ parser.cpp
@@ -41,7 +41,7 @@ int parse_str(parser_option_t *option, char *arg) {
 int parse_int(parser_option_t *option, char *arg) {
     errno = 0;
     char *end;
-    int res = strtol(arg, &end, 0);
+    long res = strtol(arg, &end, 0);
 
     if (arg[0] == '\0' || end[0] != '\0') {
         ERROR("Argument for ""%s"" is not a valid integer value",  option->long_name);
